# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Finance Mini App
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface Transaction {
  id: number;
  user_id: number;
  type: TransactionType;
  amount: number;
  category: string;
  /** @nullable */
  note?: string | null;
  created_at: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface TransactionInput {
  user_id: number;
  type: TransactionInputType;
  amount: number;
  category: string;
  note?: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface BalanceSummary {
  balance: number;
  total_income: number;
  total_expense: number;
}

export interface CategoryItem {
  category: string;
  total: number;
}

export interface CategorySummary {
  expenses: CategoryItem[];
  incomes: CategoryItem[];
}

export interface MonthlySummary {
  income: number;
  expense: number;
  net: number;
}

export interface Limit {
  id: number;
  user_id: number;
  category: string;
  monthly_limit: number;
}

export interface LimitInput {
  user_id: number;
  category: string;
  monthly_limit: number;
}

export type ListTransactionsParams = {
user_id: number;
limit?: number;
};

export type GetBalanceParams = {
user_id: number;
};

export type GetCategorySummaryParams = {
user_id: number;
};

export type GetMonthlySummaryParams = {
user_id: number;
};

export type GetLimitsParams = {
user_id: number;
};