# lib/api-zod/src/generated/types/limit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Finance Mini App
 * OpenAPI spec version: 0.1.0
 */

export interface Limit {
  id: number;
  user_id: number;
  category: string;
  monthly_limit: number;
}